+++ module-apps/application-phonebook/widgets/InputBoxWithLabelAndIconWidget.cpp
@@ -1,4 +1,4 @@
-#include "InputBoxWithLabelAndIconItem.hpp"
+#include "InputBoxWithLabelAndIconWidget.hpp"
 
 #include "application-phonebook/data/PhonebookStyle.hpp"
 
@@ -9,27 +9,28 @@
 
 namespace gui
 {
-    InputBoxWithLabelAndIconItem::InputBoxWithLabelAndIconItem(phonebookInternals::ListItemName listItemName,
-                                                               std::function<void(const UTF8 &)> bottomBarTemporaryMode,
-                                                               std::function<void()> bottomBarRestoreFromTemporaryMode)
+    InputBoxWithLabelAndIconWidget::InputBoxWithLabelAndIconWidget(
+        phonebookInternals::ListItemName listItemName,
+        std::function<void(const UTF8 &)> bottomBarTemporaryMode,
+        std::function<void()> bottomBarRestoreFromTemporaryMode)
         : listItemName(listItemName), bottomBarTemporaryMode(std::move(bottomBarTemporaryMode)),
           bottomBarRestoreFromTemporaryMode(std::move(bottomBarRestoreFromTemporaryMode))
     {
-        setMinimumSize(phonebookStyle::inputBoxWithLabelAndIconItem::w,
-                       phonebookStyle::inputBoxWithLabelAndIconItem::h);
+        setMinimumSize(phonebookStyle::inputBoxWithLabelAndIconIWidget::w,
+                       phonebookStyle::inputBoxWithLabelAndIconIWidget::h);
 
         setMargins(gui::Margins(0, style::margins::big, 0, 0));
 
-        hBox = new gui::HBox(this, 0, 0, phonebookStyle::inputBoxWithLabelAndIconItem::w, 0);
+        hBox = new gui::HBox(this, 0, 0, phonebookStyle::inputBoxWithLabelAndIconIWidget::w, 0);
         hBox->setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_NO_EDGES);
         hBox->setPenFocusWidth(style::window::default_border_focus_w);
         hBox->setPenWidth(style::window::messages::sms_border_no_focus);
 
         inputBoxLabel = new gui::Label(hBox, 0, 0, 0, 0);
-        inputBoxLabel->setMinimumSize(phonebookStyle::inputBoxWithLabelAndIconItem::input_box_w,
-                                      phonebookStyle::inputBoxWithLabelAndIconItem::input_box_h);
+        inputBoxLabel->setMinimumSize(phonebookStyle::inputBoxWithLabelAndIconIWidget::input_box_w,
+                                      phonebookStyle::inputBoxWithLabelAndIconIWidget::input_box_h);
         inputBoxLabel->setMargins(
-            gui::Margins(0, 0, phonebookStyle::inputBoxWithLabelAndIconItem::input_box_right_margin, 0));
+            gui::Margins(0, 0, phonebookStyle::inputBoxWithLabelAndIconIWidget::input_box_right_margin, 0));
         inputBoxLabel->setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_BOTTOM);
         inputBoxLabel->setAlignment(Alignment(gui::Alignment::Horizontal::Center, gui::Alignment::Vertical::Center));
         inputBoxLabel->setFont(style::window::font::medium);
@@ -38,18 +39,18 @@ namespace gui
         tickImage = new gui::Image(hBox, 0, 0, 0, 0);
         tickImage->setAlignment(Alignment(gui::Alignment::Vertical::Center));
         // Not ideal -> best solution would be to create separate widget with image inside box.
-        tickImage->setMargins(gui::Margins(phonebookStyle::inputBoxWithLabelAndIconItem::tick_image_left_margin,
+        tickImage->setMargins(gui::Margins(phonebookStyle::inputBoxWithLabelAndIconIWidget::tick_image_left_margin,
                                            0,
-                                           phonebookStyle::inputBoxWithLabelAndIconItem::tick_image_right_margin,
+                                           phonebookStyle::inputBoxWithLabelAndIconIWidget::tick_image_right_margin,
                                            0));
         tickImage->setVisible(false);
         tickImage->activeItem = false;
 
         descriptionLabel = new gui::Label(hBox, 0, 0, 0, 0);
-        descriptionLabel->setMinimumSize(phonebookStyle::inputBoxWithLabelAndIconItem::description_label_w,
-                                         phonebookStyle::inputBoxWithLabelAndIconItem::description_label_h);
+        descriptionLabel->setMinimumSize(phonebookStyle::inputBoxWithLabelAndIconIWidget::description_label_w,
+                                         phonebookStyle::inputBoxWithLabelAndIconIWidget::description_label_h);
         descriptionLabel->setMargins(
-            gui::Margins(0, 0, phonebookStyle::inputBoxWithLabelAndIconItem::description_label_right_margin, 0));
+            gui::Margins(0, 0, phonebookStyle::inputBoxWithLabelAndIconIWidget::description_label_right_margin, 0));
         descriptionLabel->setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_NO_EDGES);
         descriptionLabel->setAlignment(Alignment(gui::Alignment::Horizontal::Left, gui::Alignment::Vertical::Center));
         descriptionLabel->setFont(style::window::font::medium);
@@ -64,7 +65,8 @@ namespace gui
         setEdges(gui::RectangleEdgeFlags::GUI_RECT_EDGE_NO_EDGES);
     }
 
-    auto InputBoxWithLabelAndIconItem::onDimensionChanged(const BoundingBox &oldDim, const BoundingBox &newDim) -> bool
+    auto InputBoxWithLabelAndIconWidget::onDimensionChanged(const BoundingBox &oldDim, const BoundingBox &newDim)
+        -> bool
     {
         hBox->setPosition(0, 0);
         hBox->setSize(newDim.w, newDim.h);
@@ -72,7 +74,7 @@ namespace gui
         return true;
     }
 
-    void InputBoxWithLabelAndIconItem::applyItemNameSpecificSettings()
+    void InputBoxWithLabelAndIconWidget::applyItemNameSpecificSettings()
     {
         switch (listItemName) {
         case phonebookInternals::ListItemName::SpeedDialKey:
@@ -93,7 +95,7 @@ namespace gui
         }
     }
 
-    void InputBoxWithLabelAndIconItem::speedDialKeyHandler()
+    void InputBoxWithLabelAndIconWidget::speedDialKeyHandler()
     {
         descriptionLabel->setText(utils::localize.get("app_phonebook_new_speed_dial_key"));
         iconImage->set("small_circle");
@@ -127,7 +129,7 @@ namespace gui
         onLoadCallback = [&](std::shared_ptr<ContactRecord> contact) { inputBoxLabel->setText(contact->speeddial); };
     }
 
-    void InputBoxWithLabelAndIconItem::addToFavouritesHandler()
+    void InputBoxWithLabelAndIconWidget::addToFavouritesHandler()
     {
         descriptionLabel->setText(utils::localize.get("app_phonebook_new_add_to_fav"));
         iconImage->set("small_heart");
@@ -173,7 +175,7 @@ namespace gui
             tickImage->visible = contact->isOnFavourites();
         };
     }
-    void InputBoxWithLabelAndIconItem::addToICEHandler()
+    void InputBoxWithLabelAndIconWidget::addToICEHandler()
     {
         descriptionLabel->setText(utils::localize.get("app_phonebook_new_add_to_ice"));
         iconImage->set("ice");